requirements.txt
flask
flask-cors
openai
langdetect
deep-translator
scikit-learn
numpy
pandas
groq
python-dotenv
googletrans